+++ app/src/main/java/com/babylon/wallet/android/presentation/dapp/connectionrequest/DAppConnectionRequestViewModel.kt
@@ -1,12 +1,12 @@
-package com.babylon.wallet.android.presentation.dapp.connectionrequest.viewmodel
+package com.babylon.wallet.android.presentation.dapp.connectionrequest
 
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.babylon.wallet.android.presentation.dapp.data.DAppConnectionData
-import com.babylon.wallet.android.presentation.dapp.domain.DAppConnectionRepository
+import com.babylon.wallet.android.data.dapp.DAppConnectionData
+import com.babylon.wallet.android.domain.dapp.DAppConnectionRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
 import javax.inject.Inject
@@ -16,13 +16,13 @@ class DAppConnectionRequestViewModel @Inject constructor(
     private val dAppConnectionRepository: DAppConnectionRepository
 ) : ViewModel() {
 
-    var uiState by mutableStateOf(DAppConnectionUiState())
+    var state by mutableStateOf(DAppConnectionUiState())
         private set
 
     init {
         viewModelScope.launch {
             val dAppConnectionData = dAppConnectionRepository.getDAppConnectionData()
-            uiState = uiState.copy(
+            state = state.copy(
                 loading = false,
                 dAppConnectionData = dAppConnectionData
             )
