+++ app/src/main/java/com/babylon/wallet/android/presentation/dapp/login/ChooseDAppLoginViewModel.kt
@@ -1,12 +1,12 @@
-package com.babylon.wallet.android.presentation.dapp.login.viewmodel
+package com.babylon.wallet.android.presentation.dapp.login
 
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.babylon.wallet.android.presentation.dapp.data.DAppConnectionData
-import com.babylon.wallet.android.presentation.dapp.domain.DAppConnectionRepository
+import com.babylon.wallet.android.data.dapp.DAppConnectionData
+import com.babylon.wallet.android.domain.dapp.DAppConnectionRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
 import javax.inject.Inject
@@ -19,13 +19,13 @@ class ChooseDAppLoginViewModel @Inject constructor(
     var selected by mutableStateOf(false)
         private set
 
-    var uiState by mutableStateOf(ChooseDAppLoginUiState())
+    var state by mutableStateOf(ChooseDAppLoginUiState())
         private set
 
     init {
         viewModelScope.launch {
             val dAppConnectionData = dAppConnectionRepository.getChooseDAppLoginData()
-            uiState = uiState.copy(
+            state = state.copy(
                 loading = false,
                 dAppData = dAppConnectionData
             )
